requirements.txt
pprintpp==0.4.0
datetime
requests
pyaudio
SpeechRecognition
openai
python-dotenv
datetime
pytz
pyttsx3
selenium==3.141.0
requests==2.25.0
geopy
geocoder
psutil
wikipedia
urllib3==1.26.2
pyjokes
pyautogui
pywhatkit
wolframalpha
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0